+++ saleor/graphql/order/mutations/order_update_shipping.py
@@ -7,6 +7,7 @@ from ....order.error_codes import OrderErrorCode
 from ....permission.enums import OrderPermissions
 from ....shipping import models as shipping_models
 from ....shipping.utils import convert_to_shipping_method_data
+from ....webhook.deprecated_event_types import WebhookEventType
 from ....webhook.event_types import WebhookEventAsyncType
 from ...core import ResolveInfo
 from ...core.context import SyncWebhookControlContext
@@ -74,8 +75,15 @@ class OrderUpdateShipping(
                 "lines", "channel__shipping_method_listings"
             ),
         )
+
+        # Below this Order is properly typed Order, instead of Model
+        if not isinstance(order, models.Order):
+            raise Exception("Received node is not Order")
+
         cls.check_channel_permissions(info, [order.channel_id])
         cls.validate_order(order)
+
+        is_draft_order = order.is_draft()
 
         if "shipping_method" not in input:
             raise ValidationError(
@@ -88,7 +96,7 @@ class OrderUpdateShipping(
             )
 
         if not input.get("shipping_method"):
-            if not order.is_draft() and order.is_shipping_required():
+            if not is_draft_order and order.is_shipping_required():
                 raise ValidationError(
                     {
                         "shipping_method": ValidationError(
@@ -137,5 +145,13 @@ class OrderUpdateShipping(
 
         order.save(update_fields=SHIPPING_METHOD_UPDATE_FIELDS)
         # Post-process the results
-        call_order_event(manager, WebhookEventAsyncType.ORDER_UPDATED, order)
+
+        event_to_emit = (
+            WebhookEventAsyncType.DRAFT_ORDER_UPDATED
+            if is_draft_order
+            else WebhookEventType.ORDER_UPDATED
+        )
+
+        call_order_event(manager, event_to_emit, order)
+
         return OrderUpdateShipping(order=SyncWebhookControlContext(order))